Have the widgets try as much as possible to use the default ephemeris.
Also move the current chart variable onto the service, since we will need that once the trigger checking is integrated into the UI.

Finally, also make use of SunsetSunriseInfo's quick hour number calculation since that's faster than traversing through a list to check for the hours. As a consequence, the planetary hours widget should reconstruct the planetary hours from the ssi info if it's reading it off the service.