Make sure unit information is exposed in prefs
Fix issue #9 by making sure the user knows what units and +/-values to
put in the dialog boxes in the preferences.

https://developer.android.com/reference/androidx/preference/EditTextPreference

Previously this information was buried in the summary property for each
preference, but to the user they only see the currently entered value.

https://developer.android.com/reference/android/preference/DialogPreference#attr_android:dialogMessage

So instead, we're also making sure the information is present in the
dialog box.